requirements.txt
simple_websocket_server
websocket-client
picamera
RPi.GPIO
sox